# swagger_server/models/courier_item.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class CourierItem(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, courier_id: int=None, courier_type: str=None, regions: List[int]=None, working_hours: List[str]=None):  # noqa: E501
        """CourierItem - a model defined in Swagger

        :param courier_id: The courier_id of this CourierItem.  # noqa: E501
        :type courier_id: int
        :param courier_type: The courier_type of this CourierItem.  # noqa: E501
        :type courier_type: str
        :param regions: The regions of this CourierItem.  # noqa: E501
        :type regions: List[int]
        :param working_hours: The working_hours of this CourierItem.  # noqa: E501
        :type working_hours: List[str]
        """
        self.swagger_types = {
            'courier_id': int,
            'courier_type': str,
            'regions': List[int],
            'working_hours': List[str]
        }

        self.attribute_map = {
            'courier_id': 'courier_id',
            'courier_type': 'courier_type',
            'regions': 'regions',
            'working_hours': 'working_hours'
        }
        self._courier_id = courier_id
        self._courier_type = courier_type
        self._regions = regions
        self._working_hours = working_hours

    @classmethod
    def from_dict(cls, dikt) -> 'CourierItem':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The CourierItem of this CourierItem.  # noqa: E501
        :rtype: CourierItem
        """
        return util.deserialize_model(dikt, cls)

    @property
    def courier_id(self) -> int:
        """Gets the courier_id of this CourierItem.


        :return: The courier_id of this CourierItem.
        :rtype: int
        """
        return self._courier_id

    @courier_id.setter
    def courier_id(self, courier_id: int):
        """Sets the courier_id of this CourierItem.


        :param courier_id: The courier_id of this CourierItem.
        :type courier_id: int
        """
        if courier_id is None:
            raise ValueError("Invalid value for `courier_id`, must not be `None`")  # noqa: E501

        self._courier_id = courier_id

    @property
    def courier_type(self) -> str:
        """Gets the courier_type of this CourierItem.


        :return: The courier_type of this CourierItem.
        :rtype: str
        """
        return self._courier_type

    @courier_type.setter
    def courier_type(self, courier_type: str):
        """Sets the courier_type of this CourierItem.


        :param courier_type: The courier_type of this CourierItem.
        :type courier_type: str
        """
        allowed_values = ["foot", "bike", "car"]  # noqa: E501
        if courier_type not in allowed_values:
            raise ValueError(
                "Invalid value for `courier_type` ({0}), must be one of {1}"
                .format(courier_type, allowed_values)
            )

        self._courier_type = courier_type

    @property
    def regions(self) -> List[int]:
        """Gets the regions of this CourierItem.


        :return: The regions of this CourierItem.
        :rtype: List[int]
        """
        return self._regions

    @regions.setter
    def regions(self, regions: List[int]):
        """Sets the regions of this CourierItem.


        :param regions: The regions of this CourierItem.
        :type regions: List[int]
        """
        if regions is None:
            raise ValueError("Invalid value for `regions`, must not be `None`")  # noqa: E501

        self._regions = regions

    @property
    def working_hours(self) -> List[str]:
        """Gets the working_hours of this CourierItem.


        :return: The working_hours of this CourierItem.
        :rtype: List[str]
        """
        return self._working_hours

    @working_hours.setter
    def working_hours(self, working_hours: List[str]):
        """Sets the working_hours of this CourierItem.


        :param working_hours: The working_hours of this CourierItem.
        :type working_hours: List[str]
        """
        if working_hours is None:
            raise ValueError("Invalid value for `working_hours`, must not be `None`")  # noqa: E501

        self._working_hours = working_hours