# swagger_server/models/orders_assign_post_request.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class OrdersAssignPostRequest(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, courier_id: int=None):  # noqa: E501
        """OrdersAssignPostRequest - a model defined in Swagger

        :param courier_id: The courier_id of this OrdersAssignPostRequest.  # noqa: E501
        :type courier_id: int
        """
        self.swagger_types = {
            'courier_id': int
        }

        self.attribute_map = {
            'courier_id': 'courier_id'
        }
        self._courier_id = courier_id

    @classmethod
    def from_dict(cls, dikt) -> 'OrdersAssignPostRequest':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The OrdersAssignPostRequest of this OrdersAssignPostRequest.  # noqa: E501
        :rtype: OrdersAssignPostRequest
        """
        return util.deserialize_model(dikt, cls)

    @property
    def courier_id(self) -> int:
        """Gets the courier_id of this OrdersAssignPostRequest.


        :return: The courier_id of this OrdersAssignPostRequest.
        :rtype: int
        """
        return self._courier_id

    @courier_id.setter
    def courier_id(self, courier_id: int):
        """Sets the courier_id of this OrdersAssignPostRequest.


        :param courier_id: The courier_id of this OrdersAssignPostRequest.
        :type courier_id: int
        """
        if courier_id is None:
            raise ValueError("Invalid value for `courier_id`, must not be `None`")  # noqa: E501

        self._courier_id = courier_id